requirements.txt
streamlit==1.44.0
pandas==2.2.3
google-genai==1.14.0
pydantic==2.10.6
openpyxl==3.1.2
